+++ packages/frameworks/solid/src/progress/progress-context.ts
@@ -1,0 +1,9 @@
+import { createContext } from '../create-context'
+import { type UseProgressReturn } from './use-progress'
+
+export interface ProgressContext extends UseProgressReturn {}
+
+export const [ProgressProvider, useProgressContext] = createContext<ProgressContext>({
+  hookName: 'useProgressContext',
+  providerName: '<ProgressProvider />',
+})